Shadows and expanded scene
The Bahama oriole (Icterus northropi) is a species of songbird in the New World blackbird family - Icteridae. It is endemic to the Bahamas. It is listed as critically endangered by the IUCN (CR). The taxon was originally classified as its own distinct species in 1890 by Joel Asaph Allen before it was lumped with the Cuban oriole (Icterus melanopsis), Hispaniolan oriole (Icterus dominicensis), and Puerto Rican oriole (Icterus portoricensis) into a single species by the ornithologist James Bond in his book "Birds of the West Indies".